requirements.txt
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
inquirer==3.1.3
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2023.3
readchar==4.0.5
redis==5.0.1
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
spotipy==2.23.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
XlsxWriter==3.1.2
